+++ web/src/routes/_authorized/_default.tsx
@@ -1,6 +1,6 @@
 import { createFileRoute, Outlet } from '@tanstack/react-router';
 import { Navigation } from '../../shared/components/Navigation/Navigation';
-import { VideoTutorialsModal } from '../../shared/video-tutorials/components/VideoTutorialsModal/VideoTutorialsModal';
+import { VideoTutorialsModal } from '../../shared/video-tutorials/VideoTutorialsModal';
 import { VideoTutorialsWidget } from '../../shared/video-tutorials/VideoTutorialsWidget';
 
 export const Route = createFileRoute('/_authorized/_default')({
