+++ web/src/shared/video-tutorials/VideoTutorialsModal.scss
