+++ web/src/shared/video-tutorials/components/modal/VideoList/style.scss
