+++ web/src/shared/video-tutorials/components/widget/Thumbnail/style.scss
