+++ web/src/shared/video-tutorials/components/widget/VideoCard/style.scss
