+++ web/src/shared/video-tutorials/components/widget/VideoOverlay/style.scss
